refactor(secondary-screen): promote achievement panel to a widget
Extract the RetroAchievements panel (`_buildAchievementPanel` + its
`_buildAchievementListView`) out of `_SecondaryScreenState` into a new
`const`-constructor `AchievementPanel` widget under `secondary_screen/widgets/`.

The panel is an input-driven subtree: the host passes the state snapshot, the
two view-state flags (`listView` / `celebrate`), the l10n context, and the tap
callbacks (toggle-list-view, select-achievement). The widget re-reads no state
of its own.

The two leaf builders shared with the comments page (still in the host) are
promoted to top-level functions alongside the widget so both call sites share
one copy: `buildAchievementBadge` (pure) and `buildMissablePill` (localized;
takes the fallback context + optional l10n context). Same pattern as
`buildDockIcon` in #139.

Bodies moved verbatim (only edits: state/callback parameter substitutions).
Host `secondary_screen.dart` shrinks ~400 lines. No behavior or public-API
change. Achievement *comments* page stays in the host — a separate follow-up.
